docs init, moved modal to inner components
