+++ client/src/App.jsx
@@ -4,6 +4,7 @@ import { Routes, Route } from 'react-router-dom'
 import Success from './components/success'
 import Cancel from './components/cancel'
 import TotalPage from './totalpage'
+import Docs from './components/docs'
 
 function App() {
 
@@ -18,6 +19,7 @@ function App() {
         <Route path='/' index element={<TotalPage vibeCheckout={vibeCheckout} setVibeCheckout={setVibeCheckout}/>}/>
         <Route path='/success' element={<Success />}/>
         <Route path='/cancel' element={<Cancel />}/>
+        <Route path='/docs' element={<Docs />}/>
       </Routes>
     </>
   )
