+++ client/src/components/docs.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Docs() {
+  return (
+    <div>
+      Hello
+    </div>
+  )
+}
+
+export default Docs
